+++ src/Mango.Services.ShoppingCartAPI/Models/Dto/CartHeaderDto.cs
@@ -5,6 +5,6 @@ public class CartHeaderDto
 	public int CartHeaderId { get; set; }
 	public string? UserId { get; set; }
 	public string? CouponCode { get; set; }
-	public decimal Discount { get; set; }
-	public decimal CartTotal { get; set; }
+	public double Discount { get; set; }
+	public double CartTotal { get; set; }
 }
